feat(Partner Management APIs): add new service to project
